requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.1
discord==1.0.1
discord.py==1.3.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pycodestyle==2.5.0
pylint==2.4.4
python-dotenv==0.13.0
requests==2.23.0
six==1.13.0
urllib3==1.25.9
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
